ai-tool/fastmcp_poc/requirements.txt
fastmcp
langchain
requests
pytz
python-dotenv
newsapi-python
groq
